+++ src/utils/organization/TODO_narrow.ts
@@ -1,0 +1,6 @@
+/**
+ * Organizational helper to better mark the place where the more specific type is missing
+ *
+ * @private within the repository
+ */
+export type TODO_narrow<T> = T;